Misc github flows refactors
